model/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
datasets==3.2.0
dill==0.3.8
editdistance==0.8.1
et_xmlfile==2.0.0
etils==1.11.0
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
gviz-api==1.10.0
h5py==3.12.1
huggingface-hub==0.27.1
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
jax==0.4.38
jaxlib==0.4.38
keras==3.7.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml-dtypes==0.4.1
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
protobuf==4.25.5
pyarrow==18.1.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scipy==1.14.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboard-plugin-profile==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-rocm==2.17.0
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
